Remove use of extra component for facet button
