+++ src/pages/search/selected_facets/SelectedPersonFacetsList.tsx
@@ -1,9 +1,13 @@
+import { TFunction } from 'i18next';
+import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router';
 import { PersonSearchParameter } from '../../../api/cristinApi';
 import { PersonAggregations } from '../../../types/user.types';
-import { SelectedFacetsList } from './SelectedFacetsList';
-import { SelectedPersonFacetButton } from './SelectedPersonFacetButton';
+import { getLanguageString } from '../../../utils/translation-helpers';
+import { SelectedInstitutionFacetButton } from '../registration_search/RegistrationSearchBar';
 import { getSelectedFacetsArray } from './facetHelpers';
+import { SelectedFacetButton } from './SelectedFacetButton';
+import { SelectedFacetsList } from './SelectedFacetsList';
 
 const personFacetParams: string[] = [PersonSearchParameter.Organization, PersonSearchParameter.Sector];
 
@@ -12,6 +16,7 @@ interface SelectedPersonFacetsListProps {
 }
 
 export const SelectedPersonFacetsList = ({ aggregations }: SelectedPersonFacetsListProps) => {
+  const { t } = useTranslation();
   const [searchParams] = useSearchParams();
   const selectedFacets = getSelectedFacetsArray(searchParams, personFacetParams);
 
@@ -21,9 +26,64 @@ export const SelectedPersonFacetsList = ({ aggregations }: SelectedPersonFacetsL
 
   return (
     <SelectedFacetsList>
-      {selectedFacets.map(({ param, value }) => (
-        <SelectedPersonFacetButton key={`${param}-${value}`} param={param} value={value} aggregations={aggregations} />
-      ))}
+      {selectedFacets.map(({ param, value }) => {
+        const paramContent = getParamContent(t, param);
+        const valueContent = getValueContent(t, param, value, aggregations);
+
+        return (
+          <SelectedFacetButton
+            key={`${param}${value}`}
+            param={param}
+            value={value}
+            content={
+              <>
+                {paramContent}: {valueContent}
+              </>
+            }
+          />
+        );
+      })}
     </SelectedFacetsList>
   );
+};
+
+const getParamContent = (t: TFunction, param: string) => {
+  switch (param) {
+    case PersonSearchParameter.Organization: {
+      return t('common.institution');
+    }
+    case PersonSearchParameter.Sector:
+      return t('search.sector');
+    default:
+      return param || t('common.unknown');
+  }
+};
+
+const getValueContent = (t: TFunction, param: string, value: string, aggregations?: PersonAggregations) => {
+  switch (param) {
+    case PersonSearchParameter.Organization: {
+      const institutionLabels = aggregations?.organizationFacet?.find((bucket) => bucket.key === value)?.labels;
+      const institutionName = getLanguageString(institutionLabels);
+      if (institutionName) {
+        return institutionName;
+      } else {
+        const institutionIdentifier = Number.isInteger(+value) ? `${value}.0.0.0` : value;
+        return <SelectedInstitutionFacetButton institutionIdentifier={institutionIdentifier} />;
+      }
+    }
+    case PersonSearchParameter.Sector:
+      const sectorLabels = aggregations?.sectorFacet?.find((bucket) => bucket.key === value)?.labels;
+      const sectorName = getLanguageString(sectorLabels);
+      if (sectorName) {
+        return sectorName;
+      } else {
+        if (value === 'UC') {
+          return t('basic_data.institutions.sector_values.UHI');
+        } else {
+          return t(`basic_data.institutions.sector_values.${value}`, { defaultValue: value });
+        }
+      }
+    default:
+      return value || t('common.unknown');
+  }
 };
